[#30324971] Store well location on container_associations
Ensures that whenever wells are added to plates that the container
association takes on the positional information from the well.

It also guarantees that wells will have a location and that those
locations will be unique on an individual plate.

NOTE: tests (and probably features) currently fail because they violate
one of the constraints.