Return 404 if uploaded manifest missing
ActiveRecord::RecordNotFound gets rescued by rails automatically
and handled as a 404.
This handles situations in which users try and manually enter a URL
to retrieve a manifest which hasn't been uploaded yet; I don't
believe we generate links ourselves. The example that propmted
this certainly didn't have a referal header.

Without this we raise a 503, and the user gets a 'Something went wrong'.