Generate asset links Asynchronously
acts-as-dag performance drops as the number of ancestors increases.
This can beome especially problematic where extensive cherrypicking
has occured, as intermediate plates can result in large numbers of
ancestors for a stock plate.

In particular this change was triggered by a stock plate which had
305 ancestors, which hit particularly hard when the transfer
to tubes happened, taking over 170 seconds, resulting in timeouts.

And.... I've just realised that this isn't going to work.
We depend on asset links for finding the stock plates. We need this
almost instantly.